NO-ISSUE - Fix cluster and infra_env object cleanup (#1461)

+++ src/assisted_test_infra/test_infra/helper_classes/cluster.py
@@ -127,11 +127,15 @@ class Cluster(Entity):
         return cluster.id
 
     def delete(self):
-        self.api_client.delete_cluster(self.id)
+        self.deregister_infraenv()
+        if self.id:
+            self.api_client.delete_cluster(self.id)
+            self._config.cluster_id = None
 
     def deregister_infraenv(self):
         if self._infra_env:
             self._infra_env.deregister()
+        self._infra_env = None
 
     def get_details(self):
         return self.api_client.cluster_get(self.id)
